LayoutTests/fast/regions/region-event-expected.txt

Tests regionLayoutEvent

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

 
Flow content
PASS event.target.id is "region_1"
PASS event.target.id is "region_2"
PASS successfullyParsed is true

TEST COMPLETE
